+++ plugin-server/src/cdp/consumers/cdp-events.consumer.ts
@@ -8,7 +8,7 @@ import { Hub, RawClickHouseEvent } from '../../types'
 import { parseJSON } from '../../utils/json-parse'
 import { logger } from '../../utils/logger'
 import { HogWatcherState } from '../services/hog-watcher.service'
-import { CyclotronJobQueue } from '../services/job-queue'
+import { CyclotronJobQueue } from '../services/job-queue/job-queue'
 import { HogFunctionInvocation, HogFunctionInvocationGlobals, HogFunctionTypeType } from '../types'
 import { convertToHogFunctionInvocationGlobals } from '../utils'
 import { CdpConsumerBase } from './cdp-base.consumer'
